Keycode add
Function key add
layout change method bug fix

+++ common/keymap.h
@@ -29,6 +29,6 @@ uint16_t keymap_get_keycode(uint8_t layer, uint8_t row, uint8_t col);
 uint8_t keymap_fn_layer(uint8_t fn_bits);
 
 /* keycode to send when release Fn key without using */
-uint16_t keymap_fn_keycode(uint8_t fn_bits);
+uint8_t keymap_fn_keycode(uint8_t fn_bits);
 
 #endif
